+++ src/lib/Scenes/Home/Components/ForYou/Components/MarketingHeader.tsx
@@ -29,10 +29,7 @@ export class MarketingHeader extends Component {
         </BorderBox>
 
         <Box my={2}>
-          <Sans size="3">
-            Buying art on Artsy is easier than ever before. Our most in-demand works are now available for instant
-            purchase, with simple checkout and hassle-free shipping.
-          </Sans>
+          <Sans size="3">In-demand artworks, available to buy now</Sans>
         </Box>
 
         <InvertedButton